styles: move styles to styles folder
